Currency Related Changes

+++ service-loans/src/main/java/com/capitaworld/service/loans/domain/fundprovider/HomeLoanParameter.java
@@ -17,7 +17,7 @@ import javax.persistence.Table;
 public class HomeLoanParameter extends ProductMaster implements Serializable {
 	private static final long serialVersionUID = 1L;
 
-	private Double currency;
+	private Integer currency;
 	
 	// Net Monthly Income Range (Rs.)
 	@Column(name = "min_net_monthly_income_range")
@@ -278,11 +278,11 @@ public class HomeLoanParameter extends ProductMaster implements Serializable {
 	@Column(name = "ltv_for_eligibility")
 	private Integer ltvForEligibility;
 
-	public Double getCurrency() {
+	public Integer getCurrency() {
 		return currency;
 	}
 
-	public void setCurrency(Double currency) {
+	public void setCurrency(Integer currency) {
 		this.currency = currency;
 	}
